+++ lib/drip.dart
@@ -1,2 +1,3 @@
 export 'src/drip_core.dart';
+export 'src/drip_observer.dart';
 export 'src/widgets/widgets.dart';